refactor(logger): break logging formats apart to improve extensibility
