+++ packages/logger/src/format/constants.ts
@@ -1,0 +1,9 @@
+export const ERROR_PARAM_POSITION = 0;
+
+export const LEVEL_EMOJI: Record<string, string> = {
+  info: '📚',
+  warn: '⚠️',
+  error: '🔥',
+  debug: '🐛',
+  default: '🤷',
+};
